Fixed the button bug finally
The issue was, that instead of removing buttons from the scene, I was just hiding them. After removing them from the scene when the player, Restarts, goes back to level select, or advances to the next level prevented the buildup of "ghost" buttons that hung out a long time after the player had completed the level. This should also hopefully solve some of my memory issues.